# Android: Baseline profile generation