--- a/Source/Android/app/build.gradle
+++ b/Source/Android/app/build.gradle
@@ -33,6 +33,8 @@ android {
         versionCode(getBuildVersionCode())
 
         versionName "${getVersion()}"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     signingConfigs {
@@ -68,6 +70,19 @@ android {
             applicationIdSuffix ".debug"
             versionNameSuffix '-debug'
             jniDebuggable true
+        }
+
+        benchmark {
+            signingConfig signingConfigs.debug
+            matchingFallbacks = ['release']
+            debuggable false
+            applicationIdSuffix ".benchmark"
+            versionNameSuffix '-benchmark'
+            proguardFiles getDefaultProguardFile(
+                    'proguard-android-optimize.txt'),
+                    'proguard-rules.pro'
+            minifyEnabled true
+            shrinkResources true
         }
     }
 
@@ -107,6 +122,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.7.0'
     implementation 'androidx.core:core-splashscreen:1.0.0'
     implementation 'androidx.preference:preference:1.2.0'
+    implementation 'androidx.profileinstaller:profileinstaller:1.2.1'
 
     // Force dependency version to solve build conflict with androidx preferences
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1"