--- a/Source/Android/benchmark/src/main/AndroidManifest.xml
+++ b/Source/Android/benchmark/src/main/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <queries>
+        <package android:name="org.dolphinemu.dolphinemu.debug" />
+    </queries>
+</manifest>