--- a/Source/Android/settings.gradle
+++ b/Source/Android/settings.gradle
@@ -1,1 +1,2 @@
 include ':app'
+include ':benchmark'