@JosJuice 2022-12-10
Kind of ambivalent about adding 4000 lines to shave off 100 ms, but if others think that 100 ms improvement is important enough, I wouldn't be against merging this.