@t895 2022-12-10
> Kind of ambivalent about adding 4000 lines to shave off 100 ms, but if others think that 100 ms improvement is important enough, I wouldn't be against merging this.

It seems a little silly but as the app uses more UI libraries in the future, (Jetpack Compose) the gains will increase