@JosJuice 2022-12-10
So... That means that Jetpack Compose adds a lot of additional startup time? If so, what are the advantages to Jetpack Compose that would make it worth using? (Not very familiar with it)